Extract model account_bank_statement
