+++ account_constraints/account_constraints.py
@@ -195,64 +195,3 @@ class AccountMoveLine(orm.Model):
          "the same than the company one.",
          ['currency_id']),
     ]
-
-
-class AccountInvoice(orm.Model):
-    _inherit = "account.invoice"
-
-    def action_cancel(self, cr, uid, ids, context=None):
-        """Override the method to add the key 'from_parent_object' in
-        the context. This is to allow to delete move line related to
-        invoice through the cancel button.
-        """
-        if context is None:
-            context = {}
-        else:
-            context = context.copy()
-        context['from_parent_object'] = True
-        return super(AccountInvoice, self).action_cancel(cr, uid, ids,
-                                                         context=context)
-
-    def action_move_create(self, cr, uid, ids, context=None):
-        """Override the method to add the key 'from_parent_object' in
-        the context."""
-        if context is None:
-            context = {}
-        else:
-            context = context.copy()
-        context['from_parent_object'] = True
-        return super(AccountInvoice, self).action_move_create(cr, uid, ids,
-                                                              context=context)
-
-
-class AccountBankStatement(orm.Model):
-    _inherit = "account.bank.statement"
-
-    def button_cancel(self, cr, uid, ids, context=None):
-        """Override the method to add the key 'from_parent_object' in
-        the context. This is to allow to delete move line related to
-        bank statement through the cancel button.
-        """
-        if context is None:
-            context = {}
-        else:
-            context = context.copy()
-        context['from_parent_object'] = True
-        return super(AccountBankStatement, self).button_cancel(cr, uid, ids,
-                                                               context=context)
-
-    def create_move_from_st_line(self, cr, uid, st_line_id,
-                                 company_currency_id,
-                                 st_line_number, context=None):
-        """Add the from_parent_object key in context in order to be able
-        to post the move.
-        """
-        if context is None:
-            context = {}
-        else:
-            context = context.copy()
-        context['from_parent_object'] = True
-        return super(AccountBankStatement, self).create_move_from_st_line(
-            cr, uid, st_line_id, company_currency_id,
-            st_line_number, context=context
-        )
