+++ account_constraints/model/__init__.py
@@ -1,3 +1,4 @@
 # -*- coding: utf-8 -*-
 
 from . import account_invoice
+from . import account_bank_statement
